requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
eval_type_backport==0.2.2
fastapi==0.115.12
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
griffe==1.7.2
gunicorn==23.0.0
h11==0.14.0
heyoo==0.1.2
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
logfire-api==3.14.0
loguru==0.7.3
MarkupSafe==3.0.2
openai==1.75.0
opentelemetry-api==1.32.1
packaging==25.0
phonenumbers==9.0.3
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-ai-slim==0.0.44
pydantic-extra-types==2.10.3
pydantic-graph==0.0.44
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.0.1
pytz==2025.2
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
zenopay==0.1.1
zipp==3.21.0
